+++ src/main/java/racingcar/service/RacingCarService.java
@@ -5,7 +5,7 @@ import racingcar.service.domain.Car;
 import racingcar.service.domain.factory.CarFactory;
 import racingcar.service.model.Record;
 import racingcar.service.dto.RoundReady;
-import racingcar.service.model.RoundResult;
+import racingcar.service.dto.RoundResult;
 import racingcar.service.strategy.RoundRule;
 
 import java.util.ArrayList;
