+++ src/main/java/racingcar/service/dto/RoundResult.java
@@ -1,5 +1,6 @@
-package racingcar.service.model;
+package racingcar.service.dto;
 
+import racingcar.service.model.Record;
 import racingcar.utils.Preconditions;
 
 import java.util.List;
