+++ src/main/java/racingcar/ui/OutputView.java
@@ -1,6 +1,6 @@
 package racingcar.ui;
 
-import racingcar.service.model.RoundResult;
+import racingcar.service.dto.RoundResult;
 
 public interface OutputView {
     void output(RoundResult roundResult);
